fix: resolve the 34 compile errors reported by CI
* HardwareProbe: drop PackageManager#getSystemFeatureLevel (derive Vulkan compute
  support from the FEATURE_VULKAN_HARDWARE_LEVEL flag itself)
* ModelConfig: invalid 1.3L literal -> Double math with an explicit toLong()
* AiService: trailing lambda was bound to NativeLlmEngine's  parameter
* ResponseStreamer: produceIn returns ReceiveChannel, not Channel
* EngineHttp: OkHttp readTimeout needs a Long
* GeminiEngine: $keyParam() inside string templates -> ${keyParam()}
* NativeLlmEngine: discover() lost its return value in a synchronized block
* ChatViewModel: probe() lives on HardwareInfo's companion, not HardwareProbe
* MarkdownView: malformed generic on removeEldestEntry
* ModelSelectorSheet: add the onCancelDownload parameter the call sites pass
* SettingsSheet: Long/Double mix in the repaint-rate maths

+++ app/src/main/java/com/example/service/engine/EngineHttp.kt
@@ -18,7 +18,7 @@ internal object EngineHttp {
     fun streaming(readTimeoutSec: Int): OkHttpClient = OkHttpClient.Builder()
         .connectTimeout(20, TimeUnit.SECONDS)
         .writeTimeout(30, TimeUnit.SECONDS)
-        .readTimeout(readTimeoutSec.coerceIn(20, 300), TimeUnit.SECONDS)
+        .readTimeout(readTimeoutSec.coerceIn(20, 300).toLong(), TimeUnit.SECONDS)
         .retryOnConnectionFailure(true)
         .connectionPool(okhttp3.ConnectionPool(4, 5, TimeUnit.MINUTES))
         .build()
